Added dtype to hidden tensor

+++ src/model/nets.py
@@ -16,7 +16,8 @@ class Encoder(nn.Module):
         # Initialize hidden with zeros
         batch_size = x.shape[1]
         h0 = torch.zeros(self.num_layers * 2, batch_size, self.hidden_dim,
-                         device=x.device)
+                         dtype=x.dtype, device=x.device)
+
         output, hn = self.rnn(x, h0)
         return output, hn
